refactor: use common sortHighlights function
